PracHask2022

Description

Programacion 2 - Practico Haskell 2022

Ejercicio 4 - Uso de combinadores básicos sobre funciones primitivas

f4longi :: [a] -> Integer

Calcula la longitud de una lista (f41)

f4longi' :: [a] -> Integer

f4longi usando composicion (f41)

f4longi'' :: [a] -> Integer

f4longi pointfree (f41)

f4cuentaE :: Eq a => a -> [a] -> Integer

Cuenta cantidad de elementos iguales al 1er. arg. en una lista (f42)

find :: (Foldable t, Eq a) => a -> t a -> Bool

Retorna bool indicando si se encuentra el elemento en la lista (f43)

double :: [a] -> [a]

Genera lista a partir de otra donde cada elemento figura dos veces (f44)

double' :: [a] -> [a]

doule con ++ (f44)

double' :: [a] -> [a]

doule con ++ (f44)

greater :: (Ord a, Num p) => a -> a -> p

Retorna 1 si xs > xy, 2 si xs < xy, 0 si xs = xy (f45)

prod :: (Foldable t, Num a) => t a -> a

Retorna producto de átomos numéricos (f46)

sumsucc :: [Int] -> Int

Tomando [a, b] retorna a * b con sumas sucesivas (f47)

rev :: Foldable t => t a -> [a]

Retornar reverso de lista sin reverse (f48)

revll :: [[a]] -> [[a]]

Devuelve lista de listas invertidas (f49)

Ejercicio 5 – Funciones definidas recursivamente

f5longi :: [a] -> Int

f41 con recursión (f51)

f5cuentaE :: (Num p, Eq t) => t -> [t] -> p

f42 con recursión (f51)

f5find :: Eq a => a -> [a] -> Bool

f43 con recursión (f51)

f5double :: [a] -> [a]

f44 con recursión (f51)

f5prod :: Num a => [a] -> a

f46 con recursión (f51)

f5sumsucc :: (Eq a, Num a) => [a] -> a

f47 con recursión (f51)

f5rev :: [a] -> [a]

f48 con recursión (f51)

f5revll :: [[a]] -> [[a]]

f49 con recursión (f51)

Ejercicio 6 - Funciones recursivas con acumuladores

f6longi :: Num t => [a] -> t

f4longi con acumuladores. (f61)

f6cuentaE :: (Eq t1, Num t2) => [t1] -> t1 -> t2

f4cuentaE con acumuladores. (f62)

f6double :: [a] -> [a]

double con acumuladores. (f64)

f6prod :: Num a => [a] -> a

prod con acumuladores. (f66)

f6sumsucc :: [Int] -> Int

sumsucc con acumuladores. (f67)

f6rev :: [a] -> [a]

rev con acumuladores. (f68)

f6revll :: [[a]] -> [[a]]

revll con acumuladores. (f69)

Ejercicio 7 - Uso de patrones de recursión predefinidos

f7longi :: (Foldable t, Num a1) => t a2 -> a1

f4longi con folding, no importa el orden solo consumir la lista, así que usé foldr (f71)

f7cuentaE :: (Foldable t, Eq a1, Num a2) => a1 -> t a1 -> a2

f4cuentaE con folding, no importa el orden solo consumir la lista, así que usé foldr (f72)

f7find :: (Foldable t, Eq a) => a -> t a -> Bool

find con folding, no importa el orden solo consumir la lista, así que usé foldr (f73)

f7double :: Foldable t => t a -> [a]

double con folding, la lista no requiere orden en particular, solo duplicarlos, así que usé foldr (f74)

f7sumsucc :: [Int] -> Int

sumsucc con folding, genera una lista y la suma, la suma es asociativa así que usé foldr(f77)

f7revll :: [[a]] -> [[a]]

revll con fold, en este ejercicio importa el orden ya que no quiero alterar el orden de la lista si no más bien el orden de sus elementos (listas), que sea asociativo a derecha me permite agregar elementos en la lista de ultimo a primero, esto no cambia su orden, si usase foldl se agregarian de primero a ultimo resultando en una lista con un reverse (f79)